#27 - moved button class to it's own module
